Lift the CSV parser out of the importer, and test it
parseCSV, splitCSVLines and parseCSVLine were private methods that read
nothing off the instance except hasHeaderRow, so they move to csv/parse.ts as
functions with that passed in. The importer calls them; nothing else changed.

Quoting is where a CSV parser fails quietly - a comma inside quotes, a newline
inside a field, a doubled quote meaning one - and comprehensive-test.csv is
built around exactly that. Its recorded parse keeps a multi-line field as one
row, which is the case worth pinning.

This is the shape the entangled importers need: the conversion as functions
over data, with the Obsidian class left as the shell around it.
