+++ src/formats/csv.ts
@@ -1,6 +1,7 @@
 import { BasesConfigFile, Notice, Setting, TFolder } from 'obsidian';
 import { FormatImporter } from '../format-importer';
 import { ImportContext } from '../main';
+import { CSVRow, parseCSV } from './csv/parse';
 import {
 	TemplateConfigurator,
 	TemplateConfig,
@@ -9,10 +10,6 @@ import {
 	generateFrontmatter
 } from '../template';
 import { createBaseFile } from '../base';
-
-interface CSVRow {
-	[key: string]: string;
-}
 
 export class CSVImporter extends FormatImporter {
 	private csvHeaders: string[] = [];
@@ -55,7 +52,7 @@ export class CSVImporter extends FormatImporter {
 
 		ctx.status('Parsing ' + file.name);
 		const csvContent = await file.readText();
-		const parsedData = this.parseCSV(csvContent);
+		const parsedData = parseCSV(csvContent, this.hasHeaderRow);
 
 		// Store all rows for later processing
 		if (this.csvHeaders.length === 0 && parsedData.rows.length > 0) {
@@ -127,148 +124,6 @@ export class CSVImporter extends FormatImporter {
 		}
 		return '';
 	}
-
-	private parseCSV(content: string): { headers: string[], rows: CSVRow[] } {
-		const lines = this.splitCSVLines(content);
-		if (lines.length === 0) {
-			return { headers: [], rows: [] };
-		}
-
-		let headers: string[];
-		let startIndex: number;
-
-		if (this.hasHeaderRow) {
-			// First row contains headers
-			headers = this.parseCSVLine(lines[0]);
-			startIndex = 1;
-		}
-		else {
-			// No header row - generate column names
-			const firstRowValues = this.parseCSVLine(lines[0]);
-			headers = firstRowValues.map((_, index) => `Column ${index + 1}`);
-			startIndex = 0;
-		}
-
-		const rows: CSVRow[] = [];
-
-		for (let i = startIndex; i < lines.length; i++) {
-			const values = this.parseCSVLine(lines[i]);
-			if (values.length === 0) continue; // Skip empty lines
-
-			const row: CSVRow = {};
-			for (let j = 0; j < headers.length; j++) {
-				row[headers[j]] = values[j] || '';
-			}
-			rows.push(row);
-		}
-
-		return { headers, rows };
-	}
-
-	private splitCSVLines(content: string): string[] {
-		const lines: string[] = [];
-		let currentLine = '';
-		let inQuotes = false;
-
-		for (let i = 0; i < content.length; i++) {
-			const char = content[i];
-			const nextChar = content[i + 1];
-
-			if (char === '"') {
-				currentLine += char; // Always add the quote to the line
-				if (inQuotes && nextChar === '"') {
-					// Escaped quote - add the second quote too
-					currentLine += '"';
-					i++; // Skip next quote
-				}
-				else {
-					// Toggle quote state
-					inQuotes = !inQuotes;
-				}
-			}
-			else if (char === '\n' && !inQuotes) {
-				// End of line
-				if (currentLine.trim().length > 0) {
-					lines.push(currentLine);
-				}
-				currentLine = '';
-			}
-			else if (char === '\r' && nextChar === '\n' && !inQuotes) {
-				// Windows line ending
-				if (currentLine.trim().length > 0) {
-					lines.push(currentLine);
-				}
-				currentLine = '';
-				i++; // Skip \n
-			}
-			else if (char === '\r' && !inQuotes) {
-				// Mac line ending
-				if (currentLine.trim().length > 0) {
-					lines.push(currentLine);
-				}
-				currentLine = '';
-			}
-			else {
-				currentLine += char;
-			}
-		}
-
-		// Add last line if exists
-		if (currentLine.trim().length > 0) {
-			lines.push(currentLine);
-		}
-
-		return lines;
-	}
-
-	private parseCSVLine(line: string): string[] {
-		const values: string[] = [];
-		let currentValue = '';
-		let inQuotes = false;
-		let startOfField = true;
-
-		for (let i = 0; i < line.length; i++) {
-			const char = line[i];
-			const nextChar = line[i + 1];
-
-			if (char === '"' && startOfField) {
-				// Starting a quoted field
-				inQuotes = true;
-				startOfField = false;
-			}
-			else if (char === '"' && inQuotes) {
-				if (nextChar === '"') {
-					// Escaped quote - add one quote to the value
-					currentValue += '"';
-					i++; // Skip the next quote
-				}
-				else {
-					// End of quoted field
-					inQuotes = false;
-				}
-			}
-			else if (char === ',' && !inQuotes) {
-				// End of field
-				values.push(currentValue);
-				currentValue = '';
-				startOfField = true;
-			}
-			else {
-				// Regular character or comma inside quotes
-				if (char !== ' ' || !startOfField || currentValue.length > 0) {
-					currentValue += char;
-					startOfField = false;
-				}
-			}
-		}
-
-		// Add last value
-		values.push(currentValue);
-
-		// Trim all values
-		return values.map(v => v.trim());
-	}
-
 
 	private sanitizeYAMLKey(key: string): string {
 		// Remove special characters that aren't valid in YAML keys
