+++ tests/csv/expected/comprehensive-test.json
@@ -1,0 +1,147 @@
+{
+	"headers": [
+		"Title",
+		"Category",
+		"Subcategory",
+		"Author",
+		"Date",
+		"Priority",
+		"Completed",
+		"Rating",
+		"Tags",
+		"Content",
+		"Notes"
+	],
+	"rows": [
+		{
+			"Title": "Meeting Notes",
+			"Category": "Work",
+			"Subcategory": "Meetings",
+			"Author": "John Doe",
+			"Date": "2024-01-15",
+			"Priority": "High",
+			"Completed": "true",
+			"Rating": "8.5",
+			"Tags": "meeting planning",
+			"Content": "Discussed Q1 goals and objectives. Key points:\n- Revenue targets\n- Team expansion\n- New product launch",
+			"Notes": "Follow up next week"
+		},
+		{
+			"Title": "Project Plan",
+			"Category": "Work",
+			"Subcategory": "Projects",
+			"Author": "Jane Smith",
+			"Date": "2024-01-16",
+			"Priority": "High",
+			"Completed": "false",
+			"Rating": "9",
+			"Tags": "planning strategy",
+			"Content": "# Project Overview\n\nThis is a detailed project plan with markdown content.\n\n## Goals\n1. Complete phase 1\n2. Start phase 2",
+			"Notes": "Budget approved"
+		},
+		{
+			"Title": "Note with \"quotes\"",
+			"Category": "Personal",
+			"Subcategory": "Misc",
+			"Author": "Test User",
+			"Date": "2024-01-17",
+			"Priority": "Low",
+			"Completed": "true",
+			"Rating": "7.5",
+			"Tags": "test special",
+			"Content": "Content with \"escaped quotes\" inside.",
+			"Notes": "Watch for edge cases"
+		},
+		{
+			"Title": "Simple Note",
+			"Category": "Personal",
+			"Subcategory": "Ideas",
+			"Author": "Bob Wilson",
+			"Date": "2024-01-18",
+			"Priority": "Medium",
+			"Completed": "false",
+			"Rating": "6",
+			"Tags": "ideas",
+			"Content": "Just a simple note without complications.",
+			"Notes": ""
+		},
+		{
+			"Title": "Unicode Test",
+			"Category": "Work",
+			"Subcategory": "Testing",
+			"Author": "Test User",
+			"Date": "2024-01-19",
+			"Priority": "Low",
+			"Completed": "true",
+			"Rating": "8",
+			"Tags": "test unicode",
+			"Content": "Content with unicode: 你好世界 🌟 café",
+			"Notes": ""
+		},
+		{
+			"Title": "Empty Fields Test",
+			"Category": "",
+			"Subcategory": "",
+			"Author": "",
+			"Date": "",
+			"Priority": "",
+			"Completed": "",
+			"Rating": "",
+			"Tags": "",
+			"Content": "Some content here but many empty fields",
+			"Notes": ""
+		},
+		{
+			"Title": "Numbers Test",
+			"Category": "Data",
+			"Subcategory": "Analysis",
+			"Author": "",
+			"Date": "2024-01-21",
+			"Priority": "High",
+			"Completed": "true",
+			"Rating": "10",
+			"Tags": "numbers data",
+			"Content": "Testing numeric values and dates.",
+			"Notes": "Score is 10"
+		},
+		{
+			"Title": "Boolean Test",
+			"Category": "Settings",
+			"Subcategory": "Config",
+			"Author": "",
+			"Date": "2024-01-22",
+			"Priority": "Medium",
+			"Completed": "false",
+			"Rating": "5",
+			"Tags": "config settings",
+			"Content": "Testing boolean values in various columns.",
+			"Notes": ""
+		},
+		{
+			"Title": "Multi-line Content",
+			"Category": "Work",
+			"Subcategory": "Documentation",
+			"Author": "Writer",
+			"Date": "2024-01-23",
+			"Priority": "High",
+			"Completed": "true",
+			"Rating": "9",
+			"Tags": "docs",
+			"Content": "This is a multi-line\ncontent field that spans\nmultiple lines to test\nproper CSV parsing.",
+			"Notes": "Important"
+		},
+		{
+			"Title": "Special Chars",
+			"Category": "Tech",
+			"Subcategory": "Testing",
+			"Author": "",
+			"Date": "2024-01-24",
+			"Priority": "Low",
+			"Completed": "false",
+			"Rating": "7",
+			"Tags": "test special",
+			"Content": "Content with special characters: !@#$%^&*()_+-=[]{}|;:',.<>?",
+			"Notes": "Need validation"
+		}
+	]
+}
