+++ tests/csv/expected/sample.json
@@ -1,0 +1,58 @@
+{
+	"headers": [
+		"Title",
+		"Category",
+		"Tags",
+		"Content",
+		"Date",
+		"Priority",
+		"Completed"
+	],
+	"rows": [
+		{
+			"Title": "My First Note",
+			"Category": "Work",
+			"Tags": "project planning",
+			"Content": "This is the content of my first note.",
+			"Date": "2024-01-15",
+			"Priority": "",
+			"Completed": "true"
+		},
+		{
+			"Title": "Second Note",
+			"Category": "Personal",
+			"Tags": "ideas thoughts",
+			"Content": "Another note with some content here.",
+			"Date": "2024-01-16",
+			"Priority": "Medium",
+			"Completed": "false"
+		},
+		{
+			"Title": "Note with, comma",
+			"Category": "Work",
+			"Tags": "urgent",
+			"Content": "Content with special characters.",
+			"Date": "2024-01-17",
+			"Priority": "Low",
+			"Completed": "true"
+		},
+		{
+			"Title": "Planning Document",
+			"Category": "Work",
+			"Tags": "planning strategy",
+			"Content": "Multi-line\ncontent goes\nhere.",
+			"Date": "2024-01-18",
+			"Priority": "High",
+			"Completed": "false"
+		},
+		{
+			"Title": "Simple Note",
+			"Category": "Personal",
+			"Tags": "misc",
+			"Content": "Just a simple note.",
+			"Date": "2024-01-19",
+			"Priority": "Medium",
+			"Completed": "true"
+		}
+	]
+}
